requirements.txt
viam-sdk
RPi.GPIO